# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MIT Pokemon Automation API
  description: |
    API for the MIT Pokemon Automation and Gameplay Application.
    All endpoints require authentication via X-API-Key header.
  version: 1.0.0
  contact:
    name: MIT Pokemon Support
servers:
  - url: /api
    description: API routes
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: User-specific secret key obtained from dashboard
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
    Pokemon:
      type: object
      required:
        - id
        - pokemon_id
        - name
        - level
        - types
        - sr
      properties:
        id:
          type: string
          format: uuid
          description: Owned Pokemon record ID
        pokemon_id:
          type: integer
          description: Pokemon species ID
        name:
          type: string
          description: Pokemon species name
        level:
          type: integer
          minimum: 1
          maximum: 10
          description: Pokemon level (condensed 1-10)
        types:
          type: array
          items:
            type: string
          description: Pokemon types
        sr:
          type: number
          description: Strength Rating
        selected_moves:
          type: array
          items:
            type: string
          maxItems: 4
          description: Currently selected moves
        is_active:
          type: boolean
          description: Whether this is the active Pokemon
        sprite_url:
          type: string
          format: uri
          description: Pokemon sprite image URL
    WildPokemon:
      type: object
      required:
        - pokemon_id
        - name
        - level
        - types
        - sr
      properties:
        pokemon_id:
          type: integer
        name:
          type: string
        level:
          type: integer
          minimum: 1
          maximum: 10
        types:
          type: array
          items:
            type: string
        sr:
          type: number
    Battle:
      type: object
      required:
        - id
        - status
        - player_wins
        - wild_wins
      properties:
        id:
          type: string
          format: uuid
        wild_pokemon:
          $ref: '#/components/schemas/WildPokemon'
        difficulty:
          type: string
          enum: [easy, normal, difficult]
        player_wins:
          type: integer
          minimum: 0
          maximum: 3
        wild_wins:
          type: integer
          minimum: 0
          maximum: 3
        capture_attempts:
          type: integer
        status:
          type: string
          enum: [active, player_won, wild_won, captured, fled]
        current_capture_dc:
          type: integer
          description: Current difficulty class for capture attempt
    RoundResult:
      type: object
      required:
        - round_number
        - winner
        - roll
      properties:
        round_number:
          type: integer
          minimum: 1
          maximum: 5
        player_move:
          type: string
        winner:
          type: string
          enum: [player, wild]
        roll:
          type: integer
          minimum: 1
          maximum: 20
        base_chance:
          type: number
        type_effectiveness:
          type: string
          enum: [super_effective, neutral, not_very_effective]
        had_stab:
          type: boolean
    CaptureResult:
      type: object
      required:
        - success
        - roll
        - dc
      properties:
        success:
          type: boolean
        roll:
          type: integer
          minimum: 1
          maximum: 20
        dc:
          type: integer
        fled:
          type: boolean
          description: Whether the wild Pokemon fled after failed capture
...
paths:
...
  /battle/round:
    post:
      summary: Execute a battle round
      description: Resolves a single round of combat using the specified move
      operationId: executeRound
      tags:
        - Battle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - move_id
              properties:
                move_id:
                  type: string
                  description: ID of the move to use
      responses:
        '200':
          description: Round result
          content:
            application/json:
              schema:
                type: object
                properties:
                  round:
                    $ref: '#/components/schemas/RoundResult'
                  battle:
                    $ref: '#/components/schemas/Battle'
                  battle_ended:
                    type: boolean
        '400':
          description: Bad request (e.g., no active battle, invalid move)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
  /capture:
    post:
      summary: Attempt to capture wild Pokemon
      description: |
        Attempts to capture the wild Pokemon in the current battle.
        Can only be attempted after winning at least one round.
        Failed capture counts as a round loss.
      operationId: attemptCapture
      tags:
        - Battle
      responses:
        '200':
          description: Capture attempt result
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    $ref: '#/components/schemas/CaptureResult'
                  battle:
                    $ref: '#/components/schemas/Battle'
                  captured_pokemon:
                    $ref: '#/components/schemas/Pokemon'
                    description: Only present if capture was successful
        '400':
          description: Bad request (e.g., no active battle, no round wins)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
...